docs(server): clarify lifecycle expectations
## What

- Document the server lifecycle as setup followed by one Start call and one Stop call after Start returns.
- Remove restart and repeated lifecycle test expectations that no longer describe the supported API contract.
- Keep observer registration idempotency coverage while moving the shared dynamic checker helper into the remaining server tests.

## Why

- The intended server lifecycle is one-shot shutdown, so the previous idempotent/restart-safe wording overpromised behavior that callers should not rely on.
- Aligning the README, GoDoc, tests, and agent guidance prevents future fixes from preserving unsupported retry or restart semantics.

## Testing

- `make package=server specs` - passed
- `make specs` - passed
- `make lint` - passed
- `make sec` - passed
